+++ myriadgroestl.cpp
@@ -35,7 +35,7 @@ void myriadhash(void *state, const void *input)
 }
 
 extern int scanhash_myriad(int thr_id, uint32_t *pdata, uint32_t *ptarget,
-	uint32_t max_nonce, uint32_t *hashes_done)
+	uint32_t max_nonce, uint64_t *hashes_done)
 {
 	static THREAD uint32_t *h_found = nullptr;
