+++ pentablake.cu
@@ -438,7 +438,7 @@ void pentablake_cpu_setBlock_80(int thr_id, uint32_t *pdata, const uint32_t *pta
 static volatile bool init[MAX_GPUS] = { false };
 
 extern int scanhash_pentablake(int thr_id, uint32_t *pdata, uint32_t *ptarget,
-	uint32_t max_nonce, uint32_t *hashes_done)
+	uint32_t max_nonce, uint64_t *hashes_done)
 {
 	static THREAD uint32_t *d_hash = nullptr;
