+++ x15/whirlpoolx.cpp
@@ -38,7 +38,7 @@ extern "C" void whirlxHash(void *state, const void *input)
 	memcpy(state, hash_xored, 32);
 }
 
-int scanhash_whirlpoolx(int thr_id, uint32_t *pdata, uint32_t *ptarget, uint32_t max_nonce, uint32_t *hashes_done)
+int scanhash_whirlpoolx(int thr_id, uint32_t *pdata, uint32_t *ptarget, uint32_t max_nonce, uint64_t *hashes_done)
 {
 	const uint32_t first_nonce = pdata[19];
 	uint32_t endiandata[20];
